requirements.txt
customtkinter
opencv-python
pillow
requests
